% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project1Function.R
\name{myddt2}
\alias{myddt2}
\title{project1 function}
\usage{
myddt2(df, x, y, cond, col, species)
}
\arguments{
\item{df}{data frame passed}

\item{x}{Length}

\item{y}{weight}

\item{cond}{Specified condition}

\item{col}{What variable to use as the color decider}

\item{species}{filters by given species}
}
\value{
a gg plot of length vs width
}
\description{
project1 function
}
\examples{
\dontrun {myddt2(df = ddt, x = "LENGTH", y= "WEIGHT",cond = LENGTH > 30, col = "RIVER", species = "CCATFISH")}
}
